Pricing/requirements.txt
bottle==0.12.17
certifi==2019.6.16
entsoe-py==0.2.5
numpy==1.16.4
pandas==0.24.2
python-dateutil==2.8.0
requests==2.22.0
scipy==1.3.0
smbus-cffi==0.5.1
soupsieve==1.9.1
SQLAlchemy==1.3.5
urllib3==1.25.3